requirements.txt
pre-commit==3.4.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2==2.9.8
pandas==2.1.1
